feat: deployment script
