+++ package.json
@@ -27,7 +27,7 @@
   "devDependencies": {
     "@types/chai": "^4.3.19",
     "@types/mocha": "^10.0.7",
-    "chai": "^5.1.1",
+    "chai": "^4.2.0",
     "mocha": "^10.7.3",
     "pino-pretty": "^11.2.2",
     "tsx": "^4.19.0"
