+++ sdk/onchain/hardhat.config.ts
@@ -26,6 +26,12 @@ const config: HardhatUserConfig = {
             // zkwasm verifier contracts are too large for the default hardhat contract size limit
             allowUnlimitedContractSize: true,
         },
+        local: {
+            url: "http://127.0.0.1:8545",
+            accounts: [
+                "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80",
+            ],
+        },
         spin_glass: {
             url: "https://spin-glass-sepolia.rpc.caldera.xyz/http",
             accounts: [process.env.SPIN_GLASS_DEPLOYMENT_WALLET_PRIVATE_KEY!],
