+++ sdk/onchain/ignition/modules/GameContractOPZK.ts
@@ -1,0 +1,22 @@
+import { buildModule } from "@nomicfoundation/hardhat-ignition/modules";
+import zkverifierModule from "./ZKVerifier";
+
+export default buildModule("OPZKGameModule", (m) => {
+    const { zkwasmVerifier } = m.useModule(zkverifierModule);
+
+    const gameSystem = m.contract("SpinOPZKGameContract", [zkwasmVerifier]);
+
+    const stakingAddress = m.staticCall(gameSystem, "getStakingContract");
+
+    const stakingContract = m.contractAt("StakingContract", stakingAddress);
+
+    const storageAddress = m.staticCall(gameSystem, "getStorageContract");
+
+    const storageContract = m.contractAt("GameStateStorage", storageAddress);
+
+    return {
+        gameSystem,
+        stakingContract,
+        storageContract,
+    };
+});
